improved comment parsing
