+++ src/nosql.ts
@@ -7,6 +7,7 @@ import { JsonSchemaDocumentToOpenApiOptions, PartialOpenApiSchema } from "openap
 import { GetColumnQuantifiers, RemoveNameQuantifiers, dbTypeEnds, getDbLabel, getMermaidDiagramDb } from "./utils/sharedUtils";
 import { pluginVersion } from "./utils/constants";
 import { ConvertOpenApiToDatabaseModel, dbToOpenApi, GeneratePropertyModel } from "./utils/nosqlUtils";
+import { defaultResetOpenApi } from "./utils/constants-nosql";
 
 declare const window: Customwindow;
 
@@ -113,79 +114,6 @@ Draw.loadPlugin(function(ui) {
     const sqlInputFromNOSQL = document.createElement("textarea");
     sqlInputFromNOSQL.style.height = "200px";
     sqlInputFromNOSQL.style.width = "100%";
-
-    const defaultResetOpenApi = `
-{
-  "openapi": "3.0.0",
-  "info": {
-    "title": "nosql plugin sample",
-    "version": "${pluginVersion}",
-    "x-comment": "Generated by core-types-json-schema (https://github.com/grantila/core-types-json-schema)"
-  },
-  "paths": {},
-  "components": {
-    "schemas": {
-      "WeatherForecast": {
-        "properties": {
-          "date": {
-            "title": "WeatherForecast.date",
-            "description": "@format date-time",
-            "type": "string"
-          },
-          "temperatureC": {
-            "title": "WeatherForecast.temperatureC",
-            "description": "@format int32",
-            "type": "number"
-          },
-          "temperatureF": {
-            "title": "WeatherForecast.temperatureF",
-            "description": "@format int32",
-            "type": "number"
-          },
-          "summary": {
-            "title": "WeatherForecast.summary",
-            "nullable": true,
-            "type": "string"
-          },
-          "nestedProp": {
-            "items": {
-              "title": "WeatherForecast.nestedProp.[]",
-              "type": "string"
-            },
-            "title": "WeatherForecast.nestedProp",
-            "type": "array"
-          },
-          "child": {
-            "$ref": "#/components/schemas/Child",
-            "title": "WeatherForecast.child"
-          }
-        },
-        "required": [
-          "nestedProp"
-        ],
-        "additionalProperties": false,
-        "title": "WeatherForecast",
-        "type": "object"
-      },
-      "Child": {
-        "properties": {
-          "name": {
-            "title": "Child.name",
-            "type": "string"
-          }
-        },
-        "required": [
-          "name"
-        ],
-        "additionalProperties": false,
-        "title": "Child",
-        "type": "object"
-      }
-    }
-  }
-}
-    `;
-
     sqlInputFromNOSQL.value = defaultResetOpenApi;
     mxUtils.br(divFromNOSQL);
     divFromNOSQL.appendChild(sqlInputFromNOSQL);
